Fix find_all_subobjects, add tests
