+++ eventio/search_utils.py
@@ -1,0 +1,36 @@
+def find_type(f, eventio_type):
+    o = next(f)
+    while not isinstance(o, eventio_type):
+        o = next(f)
+
+    if not isinstance(o, eventio_type):
+        raise ValueError('Type {} not found'.format(eventio_type))
+
+    return o
+
+
+def collect_toplevel_of_type(f, eventio_type):
+    classes_under_test = [
+        o for o in f
+        if isinstance(o, eventio_type)
+    ]
+    # make sure we found some
+    assert classes_under_test
+    return classes_under_test
+
+
+def find_all_subobjects(f, structure, level=0):
+    '''
+    Find all subobjects expected in structure.
+    So if you want all AdcSums, use
+    structure = [SimTelEvent, SimTelTelEvent, SimTelTelADCSum]
+    '''
+    objects = []
+    elem = structure[level]
+
+    for o in f:
+        if isinstance(o, structure[-1]):
+            objects.append(o)
+        elif isinstance(o, elem):
+            objects.extend(find_all_subobjects(o, structure, level + 1))
+    return objects
