+++ eventio/simtel/tests/test_simtel_objects.py
@@ -3,47 +3,13 @@ import pytest
 from pytest import approx
 import numpy as np
 
+from ...search_utils import (
+    find_type,
+    collect_toplevel_of_type
+)
+
 
 test_file = resource_filename('eventio', 'resources/gamma_test.simtel.gz')
-
-
-def find_type(f, eventio_type):
-    o = next(f)
-    while not isinstance(o, eventio_type):
-        o = next(f)
-
-    if not isinstance(o, eventio_type):
-        raise ValueError('Type {} not found'.format(eventio_type))
-
-    return o
-
-
-def collect_toplevel_of_type(f, eventio_type):
-    classes_under_test = [
-        o for o in f
-        if isinstance(o, eventio_type)
-    ]
-    # make sure we found some
-    assert classes_under_test
-    return classes_under_test
-
-
-def find_all_subcontainers(f, structure, level=0):
-    '''
-    Find all subcontainers expected in structure.
-    So if you want all AdcSums, use
-    structure = [SimTelEvent, SimTelTelEvent, SimTelTelADCSum]
-    '''
-    objects = []
-    elem = structure[level]
-
-    for o in f:
-        if isinstance(o, structure[-1]):
-            objects.append(o)
-        elif isinstance(o, elem):
-            objects.extend(find_all_subcontainers(f, structure, level + 1))
-    return objects
-
